requirement.txt
requests
python-dotenv
flask
flask-sqlalchemy
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
flask-restful
numpy
